Bluetooth: Expose Bluetooth Testing API
This moves Bluetooth testing header file to the include/bluetooth.
